Some code quality improvements (more to come)
